Misc. tweaks (mostly UI)
Revised [advc.103] (city investigation through spy units): Now city investigation is an active mission with a base cost of 40 espionage, 100% success chance, and the spy unit remains in the foreign city tile after the mission (not returned to the capital). I've added an XML tag that says for each active mission whether the spy should be returned.
Investigation no longer reveals the currently researched tech, per-turn income and treasury of the city owner.

Fixed an issue with [advc.131] (AI tech evaluation) that had caused the AI to overrate Sailing at the start of the game.

Help text of the "Delete Unit" button now shows whether the deletion will decrease expenses. [advc.004b]
Known issue in K-Mod code: The subtotals on the Financial Advisor screen can be off by 1 at all times due to rounding, so the Delete button might say - correctly - that unit costs will decrease by 1, while the Financial Advisor might show a decrease by 2.

When a unit that is selected by the active player dies (in combat or disbanded), the per-turn income shown on the upper left is immediately updated to reflect possible changes in expenses. [advc.001]

Revised predicted expenses shown on the found-city button. Should no longer be off by +/-1 gpt. [advc.004b]

Tweaked the AI thresholds for financial trouble [advc.110] and merged a few lines of code by Koshling from the Caveman2Cosmos mod that take into account gold reserves.
Source (SVN revision):
https://sourceforge.net/p/caveman2cosmos/code/2405/#diff-22
And added a function AI_financialTroubleMargin that says how close the AI is to financial trouble. I'm using that function to let the city AI choose gold buildings before actually getting into trouble. Based on this SVN revision by alberts2; the original code is by Afforess though.
https://sourceforge.net/p/caveman2cosmos/code/7489/#diff-5

Resources can only be discovered on terrain types that the resource can normally be placed on. E.g. no Gold on Grassland Hill Mines. [advc.rom3]
Merged from RoM: A New Dawn (Afforess)
SVN revision (by alberts2, who merged it into Caveman2Cosmos):
https://sourceforge.net/p/caveman2cosmos/code/7898/#diff-12

Avoid some potentially costly canTrain calls in CvCityAI. Based on code by alberts2 (Caveman2Cosmos):
https://sourceforge.net/p/caveman2cosmos/code/7408/#diff-5

When the AI starts with a free Worker, it moves its starting Settler (usually founding a city) before moving the Worker. Previously, the very first AI Worker turn was always wasted. [advc.001]

Added a clause to discourage teleportation of ships into lakes (e.g. at the end of a war). [advc.046]
I've never seen this happen in a game, but this bug report for the "We the People" mod suggests to me that it could happen:
https://forums.civfanatics.com/threads/we-the-people-bug-reporting.636760/page-6#post-15344271

Fix for a display bug merged from Dawn of Civilization (Leoreth). [advc.001]
Probably didn't affect AdvCiv, but might be relevant for future additions.
https://github.com/dguenms/Dawn-of-Civilization/commit/0c0c388a2a7b8cd4815cb3f81c6d72d313630505

Merged from "We the People" mod (Nightinggale):
Private copy-constructor for CvInfoBase in order to catch errors like
CvBonusInfo kBonus = GC.getBonusInfo(eBonus)
at compile time. [advc.003e]
https://github.com/We-the-People-civ4col-mod/Mod/commit/f3d9441aef231b55a80aae66150c5829d27976c6
(I had previously done this, inspired by Nightinggale's change, for some other classes that are frequently returned by reference, but not for the Info class hierarchy.)

On the tech tree, help text for those techs that the player can currently research takes the current game state into account. This affects e.g. which building obsoletions are listed. (But all effects are still represented by the icons in the tech boxes.) [advc.004]

BUG no longer logs debug and info messages to PythonErr2.log during startup, only warnings and errors (if any). This matches the default log levels set by K-Mod 1.46. [k146]
(The base game still writes uninteresting debug info to PythonErr2.log, which, I think, can't be prevented.)

Comments in source code about BBAI logging changed. [advc.mak]

Cleaned up my implementation of [advc.117,advc.121] (AI worker moves).

Refactoring in CvDLLWidgetData::parseActionHelp, CvCity::canConstruct, CvCity::canBeSelected, CvCityAI::AI_neededDefenders and a little in CvGameTextMgr::setBuildingHelpActual and CvPlot.h. [advc.003]

+++ CvGameCoreDLL/CyUnit.cpp
@@ -557,7 +557,7 @@ bool CyUnit::canBuildRoute()
 
 int /*BuildTypes*/ CyUnit::getBuildType()
 {
-	return (int) m_pUnit ? m_pUnit->getBuildType() : (int) NO_BUILD;
+	return (int)(m_pUnit ? m_pUnit->getBuildType() : NO_BUILD); // kmodx
 }
 
 int CyUnit::workRate(bool bMax)
